requirements.txt
aniso8601==10.0.1
asttokens==3.0.0
black @ git+https://github.com/psf/black@950ec38c119916868833cb9896dab324c5d8eadd
blinker==1.9.0
cachelib==0.13.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
convertapi==2.0.0
docopt==0.6.2
docx2txt==0.9
executing==2.2.0
feedwerk==1.2.0
Flask==3.1.0
Flask-Caching==2.3.1
flask-cors==6.0.0
Flask-FileUpload==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-MySQL==1.6.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
html2text==2025.4.15
icecream==2.1.4
idna==3.10
inscriptis==0.0.3.5
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==6.0.1
Markdown==3.7
MarkupSafe==3.0.2
msgspec==0.19.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysql-to-sqlite3==2.3.0
packaging==24.2
pathspec==0.12.1
pillow==11.3.0
pipreqs==0.4.13
platformdirs==4.3.7
plum-dispatch==1.7.4
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse2==1.7.1
pytokens==0.1.10
pytz==2025.2
redis==6.2.0
requests==2.32.4
shortuuid==1.0.13
simplejson==3.20.1
six==1.17.0
Spire.Doc.Free==12.12.0
SQLAlchemy==2.0.41
tabulate==0.9.0
text-unidecode==1.3
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yarg==0.1.10
